% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_inbo_dbase.R
\name{connect_inbo_dbase}
\alias{connect_inbo_dbase}
\title{Connect to an INBO database}
\usage{
connect_inbo_dbase(database_name, timeout = 2, autoconvert_utf8 = deprecated())
}
\arguments{
\item{database_name}{char Name of the INBO database you want to connect}

\item{timeout}{Time in seconds to timeout the connection attempt.
Setting a timeout of \code{Inf} or \code{NA} indicates no timeout.
Defaults to 2 seconds.}

\item{autoconvert_utf8}{this argument is not needed any more thanks to better
encoding in R for Windows and is therefore deprecated, do not use it any more}
}
\value{
odbc connection
}
\description{
Connects to an INBO database by simply providing the database's name as an
argument.
The function can only be used from within the INBO network.
}
\details{
For more information, refer to
\href{https://tutorials.inbo.be/tutorials/r_database_access/}{this
tutorial}.
}
\examples{
\dontrun{
connection <- connect_inbo_dbase("D0152_00_Flora")
connection <- connect_inbo_dbase("W0003_00_Lims")
}
}
\seealso{
Other connection:
\code{\link{dbDisconnect,OdbcConnection-method}},
\code{\link{dbFetch,OdbcResult-method}}
}
\author{
Stijn Van Hoey \email{stijnvanhoey@gmail.com}

Els Lommelen \email{els.lommelen@inbo.be}
}
\concept{connection}
