% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc_dbdisconnect.R
\name{dbDisconnect,OdbcConnection-method}
\alias{dbDisconnect,OdbcConnection-method}
\title{Close database connection}
\usage{
\S4method{dbDisconnect}{OdbcConnection}(conn, ...)
}
\arguments{
\item{conn}{A \code{\link[DBI:DBIConnection-class]{DBI::DBIConnection}} object,
as returned by \code{connect_inbo_dbase()}.}

\item{...}{Other parameters passed on to methods.}
}
\description{
This method is an adaptation to the INBO databases from the eponymous
function in the \code{odbc} package and is an implementation of the method
\code{\link[DBI:dbDisconnect]{DBI::dbDisconnect()}} defined in the \code{DBI} package.
}
\seealso{
Other connection: 
\code{\link{connect_inbo_dbase}()},
\code{\link{dbFetch,OdbcResult-method}}
}
\concept{connection}
