% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc_dbfetch.R
\name{dbFetch,OdbcResult-method}
\alias{dbFetch,OdbcResult-method}
\title{Fetch query result from database}
\usage{
\S4method{dbFetch}{OdbcResult}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from \code{\link[DBI:DBIResult-class]{DBI::DBIResult}},
created by \code{\link[DBI:dbSendQuery]{DBI::dbSendQuery()}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other arguments passed on to methods.}
}
\description{
This method is an adaptation from the eponymous function in the \code{odbc}
package and is an implementation of the method \code{dbFetch} defined in the
\code{DBI} package.  Additional to the \code{odbc} package, it replaces a cryptic
error message by an informative error message.
}
\seealso{
Other connection: 
\code{\link{connect_inbo_dbase}()},
\code{\link{dbDisconnect,OdbcConnection-method}}
}
\concept{connection}
