% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inboveg_header.R
\name{get_inboveg_header}
\alias{get_inboveg_header}
\title{Query header information from INBOVEG}
\usage{
get_inboveg_header(
  connection,
  survey_name,
  rec_type,
  multiple = FALSE,
  collect = FALSE
)
}
\arguments{
\item{connection}{\code{dbconnection} with the database 'Cydonia'
on the \code{inbo-sql07-prd} server}

\item{survey_name}{A character string or a character vector
giving the name or names of the survey(s) for which you want to extract
header information.
If missing, all surveys are returned.}

\item{rec_type}{A character vector giving the name of record type for which
you want to extract header information e.g. \code{'Classic'}, \code{'Classic-emmer'},
\code{'Classic-ketting'}, \code{'BioHab'}, \code{'ABS'}, \code{'PPA'}.
If missing, all recording types are returned.}

\item{multiple}{If TRUE, survey_name can take a character vector with
multiple survey names that must match exactly. If FALSE (the default),
survey_name must be a single character string (one survey name) that can
include wildcards to allow partial matches}

\item{collect}{If FALSE (the default), a remote \code{tbl} object is returned.
This is like a reference to the result of the query but the full result of
the query is not brought into memory. If TRUE the full result of the query is
collected (fetched) from the database and brought into memory of the working
environment.}
}
\value{
A remote \code{tbl} object (collect = FALSE) or a \code{tibble} dataframe
(collect = TRUE) with variables
\code{RecordingGivid},
\code{Name},
\code{UserReference},
\code{Observer},
\code{LocationCode},
\code{Latitude},
\code{Longitude},
\code{Area} (in m2),
\code{Length} (in cm),
\code{Width} (in cm),
\code{VagueDateType},
\code{VagueDateBegin},
\code{VagueDateEnd},
\code{SurveyId},
\code{RecTypeID}.
}
\description{
This function queries the INBOVEG database for header
information (metadata for a vegetation-recording or relevé) for one or more
surveys and the recorder type. All records, also with 'work needed'
are selected.
See the examples for how to get information for all surveys.
}
\examples{
\dontrun{
library(inbodb)
con <- connect_inbo_dbase("D0010_00_Cydonia")

# get header information from a specific survey and a specific recording type
# and collect the data
header_info <- get_inboveg_header(con, survey_name = "OudeLanden_1979",
rec_type = "Classic", collect = TRUE)

# get header information from several specific surveys
header_severalsurveys <- get_inboveg_header(con, survey_name =
c("MILKLIM_Heischraal2012", "NICHE Vlaanderen"), multiple = TRUE)

# get header information of all surveys,  don't collect the data
all_header_info <- get_inboveg_header(con)

# close the connection when done
dbDisconnect(con)
rm(con)
}
}
\seealso{
Other inboveg: 
\code{\link{get_inboveg_classification}()},
\code{\link{get_inboveg_layer_cover}()},
\code{\link{get_inboveg_layer_qualifier}()},
\code{\link{get_inboveg_ppa}()},
\code{\link{get_inboveg_qualifier}()},
\code{\link{get_inboveg_recording}()},
\code{\link{get_inboveg_relation_recording}()},
\code{\link{get_inboveg_survey}()}
}
\concept{inboveg}
