% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inboveg_layer_information.R
\name{get_inboveg_layerinfo}
\alias{get_inboveg_layerinfo}
\title{Query layer information of recordings (relevé) from INBOVEG}
\usage{
get_inboveg_layerinfo(connection, survey_name, multiple = FALSE)
}
\arguments{
\item{connection}{dbconnection with the database 'Cydonia'
on the inbo-sql07-prd server}

\item{survey_name}{A character string or a character vector, depending on
multiple parameter, giving the name or names of the survey(s) for which you
want to extract recordings information. If missing, all surveys are returned.}

\item{multiple}{If TRUE, survey_name can take a character vector with
multiple survey names that must match exactly. If FALSE (the default),
survey_name must be a single character string (one survey name) that can
include wildcards to allow partial matches}
}
\value{
A dataframe with variables Name (of the survey), RecordingGivid (unique Id),
UserReference, LayerCode, LayerDescription, CoverCode, Coverpercentage and Mean hight (cm)
}
\description{
This function queries the INBOVEG database for layer
information on recordings  for one or more surveys.
}
\examples{
\dontrun{
library(inbodb)
con <- connect_inbo_dbase("D0010_00_Cydonia")

# get the layer information from one survey
layerinfo_heischraal2012 <- get_inboveg_layerinfo(con, survey_name =
"MILKLIM_Heischraal2012")

# get all layer qualifiers from MILKLIM surveys (partial matching)
layerinfo_milkim <- get_inboveg_layerinfo(con, survey_name = "\%MILKLIM\%")

# get layer qualifiers from several specific surveys
layerinfo_severalsurveys <- get_inboveg_layerinfo(con, survey_name =
c("MILKLIM_Heischraal2012", "NICHE Vlaanderen"), multiple = TRUE)

# get all layer qualifiers of all surveys
all_layerinfo <- get_inboveg_layerinfo(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}

}
\seealso{
Other inboveg: 
\code{\link{get_inboveg_classification}()},
\code{\link{get_inboveg_header}()},
\code{\link{get_inboveg_layerqualifier}()},
\code{\link{get_inboveg_qualifiers}()},
\code{\link{get_inboveg_recordings}()},
\code{\link{get_inboveg_relation}()},
\code{\link{get_inboveg_survey}()}
}
\concept{inboveg}
