% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inboveg_layer_qualifiers.R
\name{get_inboveg_layerqualifier}
\alias{get_inboveg_layerqualifier}
\title{Query layer qualifier information of recordings (relevé) from INBOVEG}
\usage{
get_inboveg_layerqualifier(connection, survey_name, multiple = FALSE)
}
\arguments{
\item{connection}{dbconnection with the database 'Cydonia'
on the inbo-sql07-prd server}

\item{survey_name}{A character string or a character vector, depending on
multiple parameter, giving the name or names of the survey(s) for which you
want to extract recordings information. If missing, all surveys are returned.}

\item{multiple}{If TRUE, survey_name can take a character vector with
multiple survey names that must match exactly. If FALSE (the default),
survey_name must be a single character string (one survey name) that can
include wildcards to allow partial matches}
}
\value{
A dataframe with variables Name (of the survey), RecordingGivid (unique Id),
UserReference, LayerCode, LayerDescription, QualifierCode, Qualifier Description,
Elucidation and 'NotSure' in case the qualifier is doubtful, CoverCode and Cover percentage
}
\description{
This function queries the INBOVEG database for layer
qualifier information on recordings  for one or more surveys.
}
\examples{
\dontrun{
library(inbodb)
library(DBI)
library(odbc)
library(dplyr)
con <- connect_inbo_dbase("D0010_00_Cydonia")

# get the layer qualifiers from one survey
qualifiers_heischraal2012 <- get_inboveg_layerqualifier(con, survey_name =
"MILKLIM_Heischraal2012")

# get all layer qualifiers from MILKLIM surveys (partial matching)
qualifiers_milkim <- get_inboveg_layerqualifier(con, survey_name = "\%MILKLIM\%")

# get layer qualifiers from several specific surveys
qualifiers_severalsurveys <- get_inboveg_layerqualifier(con, survey_name =
c("MILKLIM_Heischraal2012", "NICHE Vlaanderen"), multiple = TRUE)

# get all layer qualifiers of all surveys
allqualifiers <- get_inboveg_layerqualifier(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}

}
\seealso{
Other inboveg: 
\code{\link{get_inboveg_classification}()},
\code{\link{get_inboveg_header}()},
\code{\link{get_inboveg_layerinfo}()},
\code{\link{get_inboveg_qualifiers}()},
\code{\link{get_inboveg_recordings}()},
\code{\link{get_inboveg_relation}()},
\code{\link{get_inboveg_survey}()}
}
\concept{inboveg}
