% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_inboveg_qualifier.R
\name{get_inboveg_qualifier}
\alias{get_inboveg_qualifier}
\title{Query qualifier information of recordings (relevé) from INBOVEG}
\usage{
get_inboveg_qualifier(
  connection,
  survey_name,
  qualifier_type,
  multiple = FALSE
)
}
\arguments{
\item{connection}{\code{dbconnection} with the database 'Cydonia'
on the \code{inbo-sql07-prd} server}

\item{survey_name}{A character string or a character vector, depending on
multiple parameter, giving the name or names of the
survey(s) for which you want to extract recordings information. If missing,
all surveys are returned.}

\item{qualifier_type}{A character vector giving the name of qualifier type
for which you want to extract  information e.g. \code{'SQ'} (site qualifier),
\code{'MQ'}
(management qualifier).
If missing, all qualifier types are returned.}

\item{multiple}{If TRUE, survey_name can take a character vector with
multiple survey names that must match exactly. If FALSE (the default),
survey_name must be a single character string (one survey name) that can
include wildcards to allow partial matches}
}
\value{
A dataframe with variables \code{RecordingGivid} (unique Id), \code{UserReference},
\code{Observer}, \code{QualifierType}, \code{QualifierCode}, \code{Description}, \verb{2nd QualifierCode},
\verb{2nd Description}, \verb{3rd QualifierCode}, \verb{3rd Description}, \code{Elucidation}, in case
qualifier is \code{'NotSure'}, \code{ParentID}, \code{QualifierResource}
}
\description{
This function queries the INBOVEG database for
qualifier information (site qualifier or management qualifier)on recordings
for one or more surveys.
}
\examples{
\dontrun{
library(inbodb)
con <- connect_inbo_dbase("D0010_00_Cydonia")

# get the qualifiers from one survey
qualifiers_heischraal2012 <- get_inboveg_qualifier(con, survey_name =
"MILKLIM_Heischraal2012")

# get all site qualifiers (SQ) from MILKLIM surveys (partial matching)
qualifiers_milkim <- get_inboveg_qualifier(con, survey_name = "\%MILKLIM\%",
qualifier_type = "SQ")

# get qualifiers from several specific surveys
qualifiers_severalsurveys <- get_inboveg_qualifier(con, survey_name =
c("MILKLIM_Heischraal2012", "NICHE Vlaanderen"), multiple = TRUE)

# get all qualifiers of all surveys
allqualifiers <- get_inboveg_qualifier(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}

}
\seealso{
Other inboveg: 
\code{\link{get_inboveg_classification}()},
\code{\link{get_inboveg_header}()},
\code{\link{get_inboveg_layer_cover}()},
\code{\link{get_inboveg_layer_qualifier}()},
\code{\link{get_inboveg_ppa}()},
\code{\link{get_inboveg_recording}()},
\code{\link{get_inboveg_relation_recording}()},
\code{\link{get_inboveg_survey}()}
}
\concept{inboveg}
