% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetnetten_observations.R
\name{get_meetnetten_observations}
\alias{get_meetnetten_observations}
\title{Query observation data from meetnetten}
\usage{
get_meetnetten_observations(
  connection,
  scheme_name = NULL,
  species_group = NULL,
  collect = FALSE
)
}
\arguments{
\item{connection}{dbconnection with the database 'S0008_00_Meetnetten'
on the inbo-sql08-prd.inbo.be server}

\item{scheme_name}{the name of the monitoring scheme for which you want to
extract visit data. Data from multiple schemes can be selected by providing
a vector with the names of the schemes.}

\item{species_group}{the name of the species group for which you want to
extract visit data. Data from multiple species groups can be selected by
providing a vector with the names of the species groups.}

\item{collect}{If FALSE (the default), a remote tbl object is returned. This
is like a reference to the result of the query but the full result of the
query is not brought into memory. If TRUE the full result of the query is
collected (fetched) from the database and brought into memory of the working
environment.}
}
\value{
A remote tbl object (collect = FALSE) or a tibble dataframe (collect
= TRUE) with variables species_group, scheme, protocol, visit_id, start_date,
location, checklist_complete, sublocation, not_counted, sample_id,
target_species, name_nl, scientific_name, sex, activity, life_stage, count,
count_type,notes, x, y
}
\description{
This function queries the meetnetten database for observation
data (standardized counts) for a specified monitoring scheme or
for all monitoring schemes within a specified species group. When no
monitoring scheme or species group is specified, the observations of all
monitoring schemes are returned.
}
\examples{
\dontrun{
library(inbodb)
con <- connect_inbo_dbase("S0008_00_Meetnetten")

# get observations for a specific monitoring scheme and collect data
observations_treefrog <- get_meetnetten_observation(con,
scheme_name = "Boomkikker", collect = TRUE)

# get observations for a specific species_group and collect data
observations_dragonflies <- get_meetnetten_observations(con,
species_group = "libellen", collect = TRUE)

# get observations for all species and do not collect data
observations_all <- get_meetnetten_observations(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}
}
\seealso{
Other meetnetten: 
\code{\link{get_meetnetten_visits}()}
}
\concept{meetnetten}
