% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetnetten_visits.R
\name{get_meetnetten_visits}
\alias{get_meetnetten_visits}
\title{Query visit data from meetnetten}
\usage{
get_meetnetten_visits(
  connection,
  scheme_name = NULL,
  species_group = NULL,
  collect = FALSE
)
}
\arguments{
\item{connection}{dbconnection with the database 'S0008_00_Meetnetten'
on the inbo-sql08-prd.inbo.be server}

\item{scheme_name}{the name of the monitoring scheme for which you want to
extract visit data.}

\item{species_group}{the name of the species group for which you want to
extract visit data.}

\item{collect}{If FALSE (the default), a remote tbl object is returned. This
is like a reference to the result of the query but the full result of the
query is not brought into memory. If TRUE the full result of the query is
collected (fetched) from the database and brought into memory of the working
environment.}
}
\value{
A remote tbl object (collect = FALSE) or a tibble dataframe (collect
= TRUE) with variables species_group, scheme, protocol, location, visit_id,
validation_status, start_date, start_time, end_date, end_time, date_created,
visit_status, for_analysis, for_targets, notes.
}
\description{
This function queries the meetnetten database for visit
data (data about the counting process) for a specified monitoring scheme or
for all monitoring schemes within a specified species group. When no
monitoring scheme or species group is specified, the visits of all monitoring
schemes are returned.
}
\examples{
\dontrun{
library(inbodb)
con <- connect_inbo_dbase("S0008_00_Meetnetten")

# get visits for a specific monitoring scheme and collect data
visits_treefrog <- get_meetnetten_visits(con, scheme_name = "Boomkikker",
collect = TRUE)

# get visits for a specific species_group and collect data
visits_dragonflies <- get_meetnetten_visits(con, species_group = "libellen",
collect = TRUE)

# get visits for all species and do not collect data
visits_all <- get_meetnetten_visits(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}
}
\seealso{
Other meetnetten: 
\code{\link{get_meetnetten_locations}()},
\code{\link{get_meetnetten_observations}()},
\code{\link{get_meetnetten_schemes}()}
}
\concept{meetnetten}
