% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc_dbGetQuery.R
\docType{methods}
\name{dbFetch,OdbcResult-method}
\alias{dbFetch,OdbcResult-method}
\title{send and fetch query to database via odbc}
\usage{
\S4method{dbFetch}{OdbcResult}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}, created by
\code{\link[=dbSendQuery]{dbSendQuery()}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other parameters passed on to methods.}
}
\description{
This method is an adaptation from the eponymous function in the odbc package and is an implementation of the virtual method \code{dbGetQuery} defined in the \code{DBI} package.
}
