% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc_dbFetch.R
\name{dbFetch,OdbcResult-method}
\alias{dbFetch,OdbcResult-method}
\title{Fetch query result from database}
\usage{
\S4method{dbFetch}{OdbcResult}(res, n = -1, ...)
}
\arguments{
\item{res}{An object inheriting from \linkS4class{DBIResult}, created by
\code{\link[DBI:dbSendQuery]{dbSendQuery()}}.}

\item{n}{maximum number of records to retrieve per fetch. Use \code{n = -1}
or \code{n = Inf}
to retrieve all pending records.  Some implementations may recognize other
special values.}

\item{...}{Other arguments passed on to methods.}
}
\description{
This method is an adaptation from the eponymous function in the \code{odbc} package and is an implementation of the method \code{dbFetch} defined in the \code{DBI} package.  Additional to the \code{odbc} package, it replaces an cryptic error message by an informative error message.  In Windows OS and if chosen encoding = "utf-8" in connect_inbo_dbase(), it moreover assures that characters are interpreted correctly (while their presentation depends on the settings of the locale).
}
