% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meetnetten_schemes.R
\name{get_meetnetten_schemes}
\alias{get_meetnetten_schemes}
\title{Overview of monitoring schemes in meetnetten database}
\usage{
get_meetnetten_schemes(connection)
}
\arguments{
\item{connection}{dbconnection with the database 'S0008_00_Meetnetten'
on the inbo-sql08-prd.inbo.be server}
}
\value{
A tibble dataframe with variables species_group, scheme and protocol.
}
\description{
This function queries the meetnetten database to give an
overview of monitoring schemes that are included.
}
\examples{
\dontrun{
library(inbodb)
con <- connect_inbo_dbase("S0008_00_Meetnetten")

# get overview of monitoring schemes in meetnetten database
meetnetten_schemes <- get_meetnetten_schemes(con)

# Close the connection when done
dbDisconnect(con)
rm(con)
}
}
\seealso{
Other meetnetten: 
\code{\link{get_meetnetten_locations}()},
\code{\link{get_meetnetten_observations}()},
\code{\link{get_meetnetten_visits}()}
}
\concept{meetnetten}
